Fix audio distortion on release.
Distortion was due to sample_times not being kept continuous when releasing, so
the parameter of the sine wave would jump discontinuously causing audio distortion.

Note: This same problem exists for pitch-bending and can be fixed by keeping the previous
floating-point input to sin() and adding to that, instead of keeping the integer sample index
and computing a new input to sine from it.